% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compute.tab.delta}
\alias{compute.tab.delta}
\title{Compute the tables with deltas for heatmaps}
\usage{
\method{compute}{tab.delta}(ist.results, ...)
}
\arguments{
\item{ist.results}{an ist.results object (needs to have
decisions, pls and signture tables)}
}
\description{
\code{compute.tab.delta()} returns the table that
generats the recovery heatmaps
}
