% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compute.tab.signatures}
\alias{compute.tab.signatures}
\title{Compute signatures in long format (after ortholog mapping)}
\usage{
\method{compute}{tab.signatures}(ist.signatures)
}
\arguments{
\item{ist.signatures}{ist.signatures object}
}
\description{
\code{compute.tab.signatures()} returns the signatures, after
mapping orthologs, in a single data.table
}
