% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{data.list.orth}
\alias{data.list.orth}
\title{Default orthology mapping from mouse/rat/human to human}
\format{
A \code{list} with names hsapiens, mmusculus and rnorvegicus,
with \code{data.table}s as provided by \code{orthIDcon}
}
\usage{
data.list.orth
}
\description{
Orthology mappings that can be plugged into
\code{ist.translator} and \code{ist.signatures} objects.
These were generated using \code{orthIDcon}.
Generated March 30th, 2020 using biomaRt.
}
\keyword{datasets}
