% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\alias{default.args.bin.train}
\alias{default.args.bin.pls}
\alias{default.args.oc.train}
\alias{default.args.oc.svm}
\alias{default.args.pheatmap}
\alias{default.config.batch}
\title{Default arguments}
\usage{
default.args.bin.train(tuneLength = 5, number = 5, repeats = 20)

default.args.bin.pls(
  validation = "CV",
  segments = 5,
  segment.type = "random",
  jackknife = FALSE
)

default.args.oc.train(
  sampling = "cross",
  nrepeat = 5,
  cross = 5,
  error.fun = function(true, pred) mean(!pred)
)

default.args.oc.svm(
  gamma = 2^(c(-10, -5, -1, 0)),
  nu = c(0.1, 0.3, 0.5, 0.7, 0.9),
  kernel = "radial"
)

default.args.pheatmap(plottype)

default.config.batch(config.file)
}
\arguments{
\item{tuneLength, number, repeats}{numeric arguments for \code{caret::train}}

\item{validation, segments, segment.type, jackknife}{arguments for
\code{pls::plsr}}

\item{sampling, nrepeat, cross, error.fun}{arguments for
\code{e1071::tune.control}}

\item{gamma, nu, kernel}{arguments for
\code{e1071::tune.svm} and \code{e1071::svm}}

\item{plottype}{character, type of plot to return defaults for
(genemap or pathwaymap)}

\item{config.file}{character, of the form \code{batchtools.ist-X.Y}, where
\code{X} is \code{sge} or \code{email} and
\code{Y} is \code{R} or \code{tmpl}}
}
\value{
\code{default.args.*} return a \code{list} with
default arguments for \code{caret::train} and \code{pls::plsr}

\code{default.config.batch()} returns a character with
the path to the desired configuration file
}
\description{
\code{default.args.bin.train()} generates proper default
arguments for \code{caret::train}, also generating an array of seeds
given the amount of folds, repetitions and parameter values

\code{default.args.bin.pls()} generates proper default
arguments for \code{pls::plsr}

\code{default.args.oc.train()} generates proper default
arguments for \code{e1071::tune.svm}

\code{default.args.oc.svm()} generates proper default
arguments for \code{e1071::tune.svm}

\code{default.args.pheatmap()} returns proper default
arguments for plotting pheatmap heatmaps

\code{default.config.batch()} returns the path of
basic configuration files for batch job submissions using \code{batchtools}.
These are stored in the internal package directory and allow working
with an SGE environment.
}
\examples{
IST::default.args.bin.train()

IST::default.args.bin.pls()

IST::default.args.oc.train()

IST::default.args.oc.svm()

IST::default.args.pheatmap("genemap")

IST::default.config.batch("batchtools.ist-sge.tmpl")

}
