% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.perf.bin}
\alias{get.perf.bin}
\title{Extract performance metrics from ist.mod.bin object}
\usage{
get.perf.bin(object)
}
\arguments{
\item{object}{an ist.mod.bin object}
}
\description{
Extract performance metrics from ist.mod.bin object
}
