% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.perf.oc}
\alias{get.perf.oc}
\title{Extract performance metrics from ist.mod.oc object}
\usage{
get.perf.oc(object)
}
\arguments{
\item{object}{an ist.mod.oc object}
}
\description{
Extract performance metrics from ist.mod.oc object
}
