% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.plstab.bin}
\alias{get.plstab.bin}
\title{Extract pls weights from model in a data.table}
\usage{
get.plstab.bin(object)
}
\arguments{
\item{object}{an ist.mod.bin object}
}
\description{
Extract pls weights from model in a data.table
}
