% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.tab.boxplot}
\alias{get.tab.boxplot}
\title{Extract table to plot boxplots}
\usage{
get.tab.boxplot(object, quote.sub = quote(TRUE))
}
\arguments{
\item{object}{an ist.results}

\item{quote.sub}{an expression enclosed by quote(), to
subset signatures (sig.id) and pathways (pathway).
If left to TRUE, it selects all signatures and pathways}
}
\description{
Extract table to plot boxplots
}
