% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.tab.genemap}
\alias{get.tab.genemap}
\title{Extract table to plot gene heatmaps}
\usage{
get.tab.genemap(
  object,
  id.path,
  sig.ids = getSignatures(x),
  long.only = TRUE,
  vars.meta.sig,
  max.genes = Inf
)
}
\arguments{
\item{object}{an ist.results}

\item{id.path}{character, pathway to represent}

\item{sig.ids}{character, signature ids to include (default: all)}

\item{long.only}{logical, whether to return the long table
(ggplot) or the wide data with row/column annotations (pheatmap)}

\item{vars.meta.sig}{character vector, column names from the signature
metadata to include in the graphical annotations. Mandatory if
\code{long.only = FALSE}}

\item{max.genes}{numeric, maximum number of genes to display in genemap
(defaults to all). Genes will be prioritised using their sum of squares}
}
\description{
Extract table to plot gene heatmaps
}
