% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.tab.pathwaymap}
\alias{get.tab.pathwaymap}
\title{Extract table to plot pathway heatmaps}
\usage{
get.tab.pathwaymap(
  object,
  id.path = getPathways(object),
  sig.ids = getSignatures(object),
  long.only = TRUE,
  vars.meta.sig,
  vars.meta.path
)
}
\arguments{
\item{object}{an ist.results}

\item{id.path}{character, pathways to represent (default: all)}

\item{sig.ids}{character, signature ids to include (default: all)}

\item{long.only}{logical, whether to return the long table
(ggplot) or the wide data with row/column annotations (pheatmap)}

\item{vars.meta.sig}{character vector, column names from the signature
metadata to include in the graphical annotations. Mandatory if
\code{long.only = FALSE}}
}
\description{
Extract table to plot pathway heatmaps
}
