% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{getTabDecisions}
\alias{getTabMetrics}
\alias{getTabPls}
\alias{getTabSignatures}
\alias{getTabDelta}
\alias{getMetaSig}
\alias{getMetaPathways}
\alias{getPathways}
\alias{getSignatures}
\alias{getGroups}
\alias{getDataMatrix}
\alias{getDataResponse}
\alias{getSampleIds}
\alias{getGeneIds}
\title{Get class slots and processed tables}
\usage{
getTabDecisions(ist.results, mod = "bin", ...)

getTabMetrics(ist.results, mod = "bin", ...)

getTabPls(ist.results, ...)

getTabSignatures(ist.results, ...)

getTabDelta(ist.results, level = "pathway", ...)

getMetaSig(object, ...)

getMetaPathways(object, mod = "bin", ...)

getPathways(object, mod = "bin", ...)

getSignatures(object, ...)

getGroups(object, ...)

getDataMatrix(object, ...)

getDataResponse(object, ...)

getSampleIds(object, ...)

getGeneIds(object, ...)
}
\arguments{
\item{ist.results}{object holding the results (must be fitted)}

\item{mod}{character, type of model (oc, bin)}

\item{object}{\code{ist.pathways}, \code{ist.results} or
\code{ist.signatures}}
}
\value{
data.table with the desired data

getPathways, getSignatures: character vector with names
}
\description{
\code{getTabDecisions()} returns the decisions of the specified
type of model \code{mod}

\code{getTabMetrics()} returns the performance
metrics of the specified type of model \code{mod}

\code{getTabPls()} returns the coefficients and
scale parameters in the PLS models in BIN

\code{getTabSignatures()} returns the
signatures in tabular format from an \code{ist.results} object

\code{getTabDelta()} returns the delta
tables of an ist.results object
at the median (reference data), gene and pathway level
(using the \code{level} argument)

\code{getMetaSig()} returns the metadata table for the
signatures

\code{getMetaPathways()} returns the metadata table for the
pathways

\code{getPathways()} returns the pathway is in an
\code{ist.pathways} or an \code{ist.results} object

\code{getSignatures()} returns the signature ids in an
\code{ist.pathways} or an \code{ist.signatures} object

\code{getGroups()} returns the sample groups in an
\code{ist.results} object, whereas \code{getGroupLevels()}
returns their levels

\code{getDataMatrix()} returns the numeric matrix with
the readouts that are used to train the models in an \code{ist.pathways}
object, or an \code{ist.results} containing it, and
\code{getDataResponse()} gives the \code{y} vector

\code{getSampleIds()} returns the sample ids, whereas
\code{getGeneIds()} returns the gene identifiers from the readouts
that were used to train the model
(\code{ist.pathways} or \code{ist.results})
}
