% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{map.pathways.to.ref}
\alias{map.pathways.to.ref}
\title{Intersect pathways to specified genes}
\usage{
map.pathways.to.ref(path.tab, id.gene, prefix)
}
\arguments{
\item{path.tab}{data.frame with the pathway ids and genes}

\item{id.gene}{character, genes to include}
}
\value{
data.frame with intersected genes
}
\description{
Intersect pathways to specified genes
}
