% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{orthology-mappers}
\alias{orthology-mappers}
\alias{orthIDdouble}
\alias{orthIDsingle}
\alias{orthIDcon}
\title{Orthology mappers}
\usage{
orthIDdouble(org.from = NULL, org.to = NULL, ...)

orthIDsingle(org = NULL, ...)

orthIDcon(org.from = NULL, org.to = NULL, proxy, ...)
}
\arguments{
\item{org.from}{character of species name
(must match species name used in Biomart)}

\item{org.to}{character of species name
(must match species name used in Biomart)}

\item{...}{further arguments passed to \code{biomaRt::useMart()}, anything
besides \code{biomart} and \code{dataset}}

\item{org}{character of species name
(must match species name used in Biomart)}

\item{proxy}{character, proxy for the database connection; leave missing
for not setting any proxy at all}
}
\value{
data.table of id conversion table
}
\description{
\code{orthIDdouble}
identifies ortholog genes using biomart homology table.
This function is based on YS's \code{orthIDcon}
from the \code{SCISSORS} repository

\code{orthIDsingle} build a trivial homolgoy
mapping in the within-species case

\code{orthIDcon} covers the general case, dispatching
the adequate function and setting proxy settings
}
\examples{
\dontrun{df.mouse2human <- orthIDdouble("mmusculus", "hsapiens")}
 
\dontrun{df.mouse <- orthIDsingle("mmusculus")}

\dontrun{
df.mouse <- orthIDcon(org.from = "mmusculus", org.to = "mmusculus")
df.mmu2hsa <- orthIDcon(org.from = "mmusculus", org.to = "hsapiens")
}

}
