% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{overlay.fc}
\alias{overlay.fc}
\title{Overlay fold changes}
\usage{
overlay.fc(mat, v.fc)
}
\arguments{
\item{mat}{matrix, samples in rows and features in columns,
must have row and colnames}

\item{v.fc}{named vector with log fold changes to apply.
These are added directly to the gene readouts.
Zeroes and missing genes imply no change.}
}
\value{
A transformed matrix with the dimensions of \code{mat}
}
\description{
\code{overlay.fc()} adds the fold changes of the
matching genes (by name) to the columns (each column is a gene) of
the input data matrix
}
