% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot-ist}
\alias{plot-ist}
\alias{plot.ist.all}
\alias{plot.ist.boxplots}
\alias{plot.ist.genemaps}
\alias{plot.ist.pathwaymaps}
\title{Plot methods}
\usage{
plot.ist.all(x, y, type = "decision", ...)

plot.ist.boxplots(
  x,
  y = head(getPathways(x), 4),
  sig.ids = c(getSignatures(x), getGroupLevels(x)),
  mapping = NULL,
  facet_rows = NULL,
  facet_cols = ggplot2::vars(pathway),
  main = NULL,
  main.width = 50,
  ...
)

plot.ist.genemaps(
  x,
  y,
  type = "ggplot",
  sig.ids = getSignatures(x),
  max.genes = Inf,
  text.size = 3,
  vars.meta.sig = NULL,
  main = y,
  main.width = 50,
  args.pheatmap = default.args.pheatmap("genemap"),
  ...
)

plot.ist.pathwaymaps(
  x,
  y = getPathways(x),
  type = "ggplot",
  sig.ids = getSignatures(x),
  text.size = 3,
  vars.meta.sig = NULL,
  vars.meta.path = NULL,
  args.pheatmap = default.args.pheatmap("pathwaymap"),
  main = NULL,
  main.width = 50,
  ...
)
}
\arguments{
\item{x, y}{for \code{plot.ist.pathwaymaps()}, respectively fitted
\code{ist.results} and a character with multiple pathways to plot}

\item{type}{character, type of plot (default "ggplot" or "pheatmap")}

\item{...}{ignored at the moment}

\item{mapping}{call to \code{aes(...)} to overlay aesthetics in the
boxplots (e.g. \code{aes(fill = org.name)})}

\item{facet_rows, facet_cols}{for \code{plot.ist.boxplots()},
facetting variables}

\item{main, main.width}{plot title (character, or NULL) and maximum width
of title wrapping (integer)}

\item{max.genes}{in \code{plot.ist.genemaps()}, numeric,
maximum number of genes to display in genemap
(defaults to all). Genes will be prioritised using their sum of squares,
useful for large heatmaps.}

\item{text.size}{numeric, size of text for heatmaps}

\item{vars.meta.sig, vars.meta.path}{for pheatmap only,
NULL not to display any signature or pathway metadata, or metadata
column names otherwise. Pathways are automatically given the \code{n.genes}
metadata, with the number of mapped genes.}

\item{args.pheatmap}{list with further arguments
to pass to \code{pheatmap()}}
}
\value{
a \code{ggplot} object (plotted as well)

\code{plot.ist.boxplots()} return a list with the plot
object and the plotted data

\code{plot.ist.genemaps()} return a list with the plot
object and the plotted data

\code{plot.ist.pathwaymaps()} return a list with the plot
object and the plotted data
}
\description{
\code{plot.ist.all()} runs on
fitted \code{ist.translator} and \code{ist.discriminator} objects

\code{plot.ist.boxplots()} displays the
pathway-wise boxplots

\code{plot.ist.genemaps()} displays the
gene heatmaps of a selected pathway

\code{plot.ist.pathwaymaps()} displays the
pathway heatmaps of a selection of pathways
}
