% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sample.data.ist}
\alias{sample.data.ist}
\title{Sample data to run IST}
\format{
A list with human expression \code{X.hsa},
human response \code{y.hsa},
murine random fold changes \code{fc.mmu.rnd},
(some of the) murine true fold changes \code{fc.mmu.true},
\code{path.hsa.oc} and \code{path.hsa.bin} for synthetic pathways
(random fixed-size ones) and,
for traceability, \code{df.orth} with
the orthology mappings within
the sample data.
For the multi-signature analysis,
the slot \code{dt.fc.mmu} contains murine signatures and
\code{df.meta.mmu} their metadata
}
\usage{
sample.data.ist
}
\description{
The sample data is a list with the
essential elements to run the complete IST.
Comes from synthetic gaussian random variables
}
\keyword{datasets}
