% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sample.env.ist}
\alias{sample.env.ist}
\title{Sample data to run IST (II)}
\format{
An environment with human expression \code{X.hsa},
human response \code{y.hsa}, list of signatures \code{list.sig}
and their metadata \code{df.meta.sig}, table of pathways
\code{df.path} and their metadata \code{df.meta.path},
and other non-essential variables (\code{de.genes} for
a list of differential limma genes and \code{list.orth} for
the orthology mappings)
}
\usage{
sample.env.ist
}
\description{
The sample data is an environment with the
essential elements to run the complete IST.
This is an updated version of \code{sample.data.ist}, also using
synthetic gaussian random variables.
}
\keyword{datasets}
