% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sample.results.ist}
\alias{sample.results.ist}
\title{Sample results object from IST run}
\format{
An object of class \code{ist.results} which implicitly
contains the \code{ist.pathways} and the \code{ist.signatures} objects
used to fit it.
}
\usage{
sample.results.ist
}
\description{
The sample output data is an \code{ist.results} object,
allows playing with plotting and interactive visualisation.
Comes from the synthetic data in \code{sample.env.ist}
}
\keyword{datasets}
