% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show-ist}
\alias{show-ist}
\alias{show.ist.translator}
\alias{show.ist.signatures}
\alias{show.ist.discriminator}
\alias{show.ist.pathways}
\alias{show.ist.results}
\title{Show methods}
\usage{
show.ist.translator(object)

show.ist.signatures(object)

show.ist.discriminator(object)

show.ist.pathways(object)

show.ist.results(object)
}
\arguments{
\item{object}{an \code{ist.translator} or \code{ist.discriminator}}
}
\value{
\code{NULL}
}
\description{
\code{show.ist.translator()} shows an overview of an
\code{ist.translator} object.

\code{show.ist.signatures()} shows an overview of an
\code{ist.signatures} object.

\code{show.ist.discriminator()} shows an overview of an
\code{ist.discriminator} object.

\code{show.ist.pathways()} shows an overview of an
\code{ist.pathways} object.

\code{show.ist.results()} shows an overview of an
\code{show.ist.results} object.
}
