% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show.logical.ids}
\alias{show.logical.ids}
\title{Show logical ids}
\usage{
show.logical.ids(object, ids, slot, prefix)
}
\description{
\code{show.logical.ids()} is a helper to summarise
logical vectors indicating samples
}
