% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{vec.ensembl2symbol}
\alias{vec.ensembl2symbol}
\alias{vec.ensembl2symbolonly}
\alias{vec.ensembl2entrez}
\title{Default labels for human ENSEMBL genes}
\format{
A named vector, directly as needed by \code{ist.results}
}
\usage{
vec.ensembl2symbol
}
\description{
These mappings (\code{vec.ensembl2symbol}, \code{vec.ensembl2symbolonly},
\code{vec.ensembl2entrez}) allow a convenient mapping
between ENSEMBL identifiers and more understandable labels
when plotting an \code{ist.results} object.
Generated March 30th, 2020 using biomaRt.
}
\keyword{datasets}
