% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrappers}
\alias{wrappers}
\alias{do.isx}
\alias{do.batch.isx}
\title{Wrapper functions}
\usage{
\method{do}{isx}(signature, object.class, name, org.from, org.to, X, y, ...)

\method{do}{batch.isx}(
  se.orgto,
  list.sig,
  object.class,
  flavour,
  mapping.data.table,
  de.genes,
  org.from,
  org.to,
  id.oc,
  id.bin,
  id.ref,
  id.ist,
  pathways.table.oc,
  pathways.table.bin,
  se.coldata.col = "ist.class",
  reg.name = "reg-isx",
  waitForJobs = TRUE,
  conf.sge = IST::default.config.batch("batchtools.ist-sge.R"),
  conf.email,
  ...
)
}
\arguments{
\item{signature, mapping.data.table}{gene signature and (optional)
precomputed orthology mapping}

\item{object.class}{character, either \code{"ist.discriminator"} to run the
classical flavours (isa/iss/isd) or \code{"ist.pathways"} to fit
pathway models}

\item{name}{character, for traceability, will be returned as-is}

\item{org.from, org.to, X, y, de.genes, id.oc, id.bin, id.ref, id.ist}{arguments
for in silico treatment}

\item{...}{\code{do.isx()}: further arguments for fitting the discriminator,
including the flavour of the ist; \code{do.batch.isx()}: ignored}

\item{se.orgto}{\code{SummarizedExperiment} object with the \code{org.to}
molecular readouts (includes the variables \code{X} and \code{y}
implicitly)}

\item{list.sig}{list with all the signatures to apply in batch;
names will be preserved in output}

\item{flavour}{character vector, flavours of ist}

\item{pathways.table.oc, pathways.table.bin}{arguments for pathway models}

\item{se.coldata.col}{character, column from the \code{SummarizedExperiment}
\code{colData()} that contains the response \code{y}}

\item{reg.name}{character, path to save the registry}

\item{waitForJobs}{logical, should the function wait until all the
jobs are processed (and then apply a reduce)? If \code{FALSE},
the function returns right after the job submission.}

\item{conf.sge}{character, path to the \code{.R} configuration file for
SGE (see \code{?makeClusterFunctionsSGE}) in batchtools}

\item{conf.email}{character, optional path to the \code{.R}
configuration file for sending an e-mail after the main job is completed}
}
\value{
\code{do.isx()}: \code{list} with the following fields:
\code{name} for traceability, and \code{decisions}

\code{do.batch.isx()}: if \code{waitForJobs = TRUE}, it returns
the reduced results, i.e. a data.frame with all the runs. Otherwise,
it returns the \code{batchtools::batchMap} result
}
\description{
\code{do.isx()} runs the ISA/ISD/ISS flavour

\code{do.batch.isx()} submits an array of jobs that iterate
over \code{do.isx()}.
If \code{object.class == "ist.discriminator"}, then \code{de.genes}
can be optionally set.
If \code{object.class == "ist.pathways"}, one should provide
\code{pathways.table.bin} and/or \code{pathways.table.bin}
}
\examples{
data(sample.data.ist)
## IST flavours
list.flavours <- with(sample.data.ist, {
do.isx(
signature = fc.mmu.true, 
object.class = "ist.discriminator", 
mapping.data.table = df.orth, 
name = "test",
flavour = c("isd", "isa", "iss"),
org.from = "mmusculus", 
org.to = "hsapiens", 
X = X.hsa, 
y = y.hsa, 
id.oc = rep(TRUE, nrow(X.hsa)), 
id.bin = rep(TRUE, nrow(X.hsa)), 
id.ref = y.hsa == 1, 
id.ist = y.hsa == -1, 
ncomp = 3
)
})
summary(list.flavours$decisions)
## IST pathways
list.pathways <- with(sample.data.ist, {
do.isx(
signature = fc.mmu.true, 
object.class = "ist.pathways", 
mapping.data.table = df.orth, 
name = "test",
pathways.table.oc = head(path.hsa.oc, 100),
pathways.table.bin = head(path.hsa.bin, 100),
org.from = "mmusculus", 
org.to = "hsapiens", 
X = X.hsa, 
y = y.hsa, 
id.oc = rep(TRUE, nrow(X.hsa)), 
id.bin = rep(TRUE, nrow(X.hsa)), 
id.ref = y.hsa == 1, 
id.ist = y.hsa == -1, 
ncomp = 3
)
})
summary(list.pathways$decisions)

}
